Move input-related components into input-folder
