+++ src/routes/username/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import PickUsername from "$lib/components/PickUsername.svelte";
+	import PickUsername from "$lib/components/Input/PickUsername.svelte";
 
 </script>
 <PickUsername></PickUsername>
